+++ Modules.php
@@ -26,45 +26,22 @@ if ( ! isset( $_REQUEST['modfunc'] ) )
 
 $_ROSARIO['page'] = 'modules';
 
-// Not printing PDF.
-if ( ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
+// Save $_REQUEST vars in session: used to recreate $_REQUEST in Bottom.php.
+if ( empty( $_REQUEST['LO_save'] )
+	&& ( mb_strpos( $modname, 'misc/' ) === false
+		|| $modname === 'misc/Portal.php'
+		|| $modname === 'misc/Registration.php'
+		|| $modname === 'misc/Export.php' )
+	&& $modname !== 'Reports/SavedReports.php' )
 {
-	// Save $_REQUEST vars in session: used to recreate $_REQUEST in Bottom.php.
-	if ( empty( $_REQUEST['LO_save'] )
-		&& ( mb_strpos( $modname, 'misc/' ) === false
-			|| $modname === 'misc/Portal.php'
-			|| $modname === 'misc/Registration.php'
-			|| $modname === 'misc/Export.php' )
-		&& $modname !== 'Reports/SavedReports.php' )
-	{
-		$_SESSION['_REQUEST_vars'] = $_REQUEST;
-	}
-
-	// Popup window detection.
-	$_ROSARIO['is_popup'] = isPopup( $modname, $_REQUEST['modfunc'] );
-
-	// AJAX request detection.
-	$_ROSARIO['not_ajax'] = empty( $_SERVER['HTTP_X_REQUESTED_WITH'] )
-		|| $_SERVER['HTTP_X_REQUESTED_WITH'] !== 'XMLHttpRequest';
-
-	// Output Header HTML.
-	if ( ( $_ROSARIO['is_popup'] && $_ROSARIO['not_ajax'] )
-		|| $_ROSARIO['not_ajax'] )
-	{
-		Warehouse( 'header' );
-	}
-	elseif ( $ETagCache )
-	{
-		// Start buffer (to generate ETag).
-		ob_start();
-	}
+	$_SESSION['_REQUEST_vars'] = $_REQUEST;
 }
-// Print PDF.
-else
-{
-	// Start buffer.
-	ob_start();
-}
+
+// Set Popup window detection.
+isPopup( $modname, $_REQUEST['modfunc'] );
+
+// Output Header HTML.
+Warehouse( 'header' );
 
 
 /**
@@ -100,12 +77,9 @@ if ( ! $allowed )
 
 				// Eg: "Student_Billing/Statements.php&_ROSARIO_PDF".
 				$_ROSARIO['ProgramLoaded'] = $program;
-			}
-		}
 
-		if ( $allowed )
-		{
-			break;
+				break 2;
+			}
 		}
 	}
 }
@@ -130,7 +104,4 @@ elseif ( User( 'USERNAME' ) )
 }
 
 // Output Footer HTML.
-if ( ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
-{
-	Warehouse( 'footer' );
-}
+Warehouse( 'footer' );